+++ oop-labs/shapes/ShapesFactory.h
@@ -3,15 +3,13 @@
 #include "LineSegment.h"
 #include "Rectangle.h"
 #include "Triangle.h"
-
 #include <memory>
-#include <vector>
 
 class ShapesFactory
 {
 public:
-	static std::shared_ptr<LineSegment> CreateLineSegment(const Point& startPosition, const Point& endPosition, uint32_t outlineColor, double outlineWidth);
-	static std::shared_ptr<Rectangle> CreateRectangle(const Point& topLeft, const Point& bottomRight, uint32_t outlineColor, double outlineWidth, uint32_t fillColor);
-	static std::shared_ptr<Triangle> CreateTriangle(const Point& vertex1, const Point& vertex2, const Point& vertex3, uint32_t outlineColor, double outlineWidth, uint32_t fillColor);
-	static std::shared_ptr<Circle> CreateCircle(const Point& center, double radius, uint32_t outlineColor, double outlineWidth, uint32_t fillColor);
+	static std::shared_ptr<LineSegment> CreateLineSegment(const Point& startPosition, const Point& endPosition, uint32_t outlineColor, float outlineWidth);
+	static std::shared_ptr<Rectangle> CreateRectangle(const Point& topLeft, const Point& bottomRight, uint32_t outlineColor, float outlineWidth, uint32_t fillColor);
+	static std::shared_ptr<Triangle> CreateTriangle(const Point& vertex1, const Point& vertex2, const Point& vertex3, uint32_t outlineColor, float outlineWidth, uint32_t fillColor);
+	static std::shared_ptr<Circle> CreateCircle(const Point& center, float radius, uint32_t outlineColor, float outlineWidth, uint32_t fillColor);
 };
